\name{yeast}
\docTye{data}

\alias{yeast}

\title{Gene expression data of the yeast cell cylce}

\description{The data contains gene expression measurements for 
  3000 randomly chosen genes of the yeast mutant cdc28 as performed and described
  by Cho \emph{et al.} For details, see the reference.}

\usage{data(yeast)}

\format{An object of class \dQuote{marrayRaw}}

\source{The data was downloaded from \emph{Yeast Cell Cylce Analysis Project} webside 
 and converted to an \emph{exprSet} object.}

\references{Cho et al., A genome-wide transcriptional analysis of the
mitotic cell cycle, Mol Cell. 1998 Jul;2(1):65-73.}
\keyword{datasets}


