\name{yeast.table2}

\docType{data}

\alias{yeast.table2}

\title{Gene expression data of the yeast cell cycle as table}

\description{The data serves as an example for the format required to
  upload tables with expression data into \emph{Mfuzzgui.} The first
  row contains the names of the samples and the first column contains
  unique identifiers for genes. To input measurement time and gene
  names, refer to \code{\link[Mfuzz]{yeast.table}}.
  
 The exemplary tables can be found in the data sub-folder of the 
 \emph{Mfuzzgui} package. 

  }

%\usage{data(yeast.table2)}


\references{Cho et al., A genome-wide transcriptional analysis of the
  mitotic cell cycle, Mol Cell. 1998 Jul;2(1):65-73.}

\seealso{\code{\link[Mfuzz]{yeast.table}} }

\keyword{datasets}


